add method for adding parameter files to patch
