+++ ImasArchiveLibTest/PatchZipTest.cs
@@ -36,12 +36,13 @@ namespace ImasArchiveLibTest
         }
 
         [DataTestMethod]
-        [DataRow("patch/patch_add.zip", "other/newfont.par", "other/text_ja_jp.xlsx", "patch/translatedcommu.xlsx", "patch/images")]
-        public async Task AddToPatchTest(string filename, string newFileName, string jajpName, string commuName, string gtfDir)
+        [DataRow("patch/patch_add.zip", "other/newfont.par", "other/text_ja_jp.xlsx", "patch/parameter_tl.xlsx", "patch/translatedcommu.xlsx", "patch/images")]
+        public async Task AddToPatchTest(string filename, string newFileName, string jajpName, string parameterName, string commuName, string gtfDir)
         {
             using PatchZipFile patchZipFile = new PatchZipFile(filename, PatchZipMode.Create);
             patchZipFile.AddFile(newFileName, "im2nx_font.par");
             patchZipFile.AddJaJp(jajpName, "text/im2nx_text.ja_jp");
+            patchZipFile.AddParameterFiles(parameterName);
             await patchZipFile.AddCommus(commuName);
             await patchZipFile.AddGtfs(gtfDir);
         }
